Validacion Unidad Medida Request

+++ SAI/app/Http/Controllers/UnidadMedidaController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Laracast\Flash\Flash;
 use App\UnidadMedida;
+use App\Http\Requests\UnidadMedidaRequest;
+use Auth;
 
 class UnidadMedidaController extends Controller
 {
@@ -27,7 +29,15 @@ class UnidadMedidaController extends Controller
      */
     public function create()
     {
-        return view('admin.producto.unidadmedida.create');
+        if (Auth::user()->rol->rol_rol === 'Administrador' || Auth::user()->rol->rol_rol === 'Encargado'){
+            return view('admin.producto.unidadmedida.create');
+        }else{
+
+            flash('Solo los usuarios con el rol "Administrador" o "Encargado" pueden registrar.')->error();
+            return redirect()->back();
+
+        }
+        
     }
 
     /**
@@ -36,13 +46,24 @@ class UnidadMedidaController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(UnidadMedidaRequest $request)
     {
-        $unidadmedida = new UnidadMedida($request->all());
-        $unidadmedida->save();
+        if (Auth::user()->rol->rol_rol === 'Administrador' || Auth::user()->rol->rol_rol === 'Encargado'){
+            $unidadmedida = new UnidadMedida($request->all());
+            $unidadmedida->save();
+            flash("Registro de la unidad de medida '' ". $request->uni_medida ." '' exitoso.")->success();
+            return redirect()->route('unidadmedida.index');
+        }else{
 
-        flash("Registro de la unidad de medida '' ". $request->uni_medida ." '' exitoso.")->success();
-        return redirect()->route('unidadmedida.index');
+            flash('Solo los usuarios con el rol "Administrador" o "Encargado" pueden registrar.')->error();
+            return redirect()->back();
+
+        }
+        
+        
+
+        
+        
     }
 
     /**
@@ -64,9 +85,23 @@ class UnidadMedidaController extends Controller
      */
     public function edit($id)
     {
-        $unidadmedida = UnidadMedida::find($id);
+        if (Auth::user()->rol->rol_rol === 'Administrador' || Auth::user()->rol->rol_rol === 'Encargado'){
+            $unidadmedida = UnidadMedida::find($id);
+            if ($unidadmedida !== null) {
+                return view('admin.producto.unidadmedida.edit')->with(compact('unidadmedida'));
+            }else{  
+                flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+                return redirect()->route('producto_articulo.index');
+            }
+        }else{
 
-        return view('admin.producto.unidadmedida.edit')->with(compact('unidadmedida'));
+            flash('Solo los usuarios con el rol "Administrador" o "Encargado" pueden modificar.')->error();
+            return redirect()->back();
+
+        }
+        
+
+        
     }
 
     /**
@@ -76,14 +111,27 @@ class UnidadMedidaController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(UnidadMedidaRequest $request, $id)
     {
-        $unidadmedida = UnidadMedida::find($id);
-        $unidadmedida->uni_medida = $request->uni_medida;
-        $unidadmedida->save();
+        if (Auth::user()->rol->rol_rol === 'Administrador' || Auth::user()->rol->rol_rol === 'Encargado'){
+            $unidadmedida = UnidadMedida::find($id);
+            if ($unidadmedida !== null) {
+                $unidadmedida->uni_medida = $request->uni_medida;
+                $unidadmedida->save();
 
-        flash("Modificación de la unidad de medida '' ". $request->uni_medida ." '' exitoso.")->success();
-        return redirect()->route('unidadmedida.index');
+                flash("Modificación de la unidad de medida '' ". $request->uni_medida ." '' exitoso.")->success();
+                return redirect()->route('unidadmedida.index');
+            }else{  
+                flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+                return redirect()->route('producto_articulo.index');
+            }
+        }else{
+
+            flash('Solo los usuarios con el rol "Administrador" o "Encargado" pueden modificar.')->error();
+            return redirect()->back();
+
+        }
+        
     }
 
     /**
@@ -94,9 +142,23 @@ class UnidadMedidaController extends Controller
      */
     public function destroy($id)
     {
-        $unidadmedida = UnidadMedida::find($id);
-        $unidadmedida->delete();
-        flash("Eliminación de la unidad de medida '' ". $unidadmedida->uni_medida ." '' exitoso.")->success();
-        return redirect()->route('unidadmedida.index');
+        if (Auth::user()->rol->rol_rol === 'Administrador'){
+            $unidadmedida = UnidadMedida::find($id);
+            if ($unidadmedida !== null) {
+                $unidadmedida->delete();
+                flash("Eliminación de la unidad de medida '' ". $unidadmedida->uni_medida ." '' exitoso.")->success();
+                return redirect()->route('unidadmedida.index');
+            }else{  
+                flash('No hay ningun registro en la Base de Datos del objeto buscado.')->error();
+                return redirect()->route('producto_articulo.index');
+            }
+        }else{
+
+            flash('Solo los usuarios con el rol "Administrador"  puede eliminar.')->error();
+            return redirect()->back();
+
+        }
+        
+        
     }
 }
